requirements.txt
absl-py==2.1.0
altair==5.4.1
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.3.0
arabic-reshaper==3.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.9.0
cachetools==5.5.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cssselect2==0.7.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Django==5.1.1
dnspython==2.7.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.115.8
fastjsonschema==2.19.1
filelock==3.16.1
Flask==3.1.0
Flask-RESTful==0.3.10
flatbuffers==24.3.25
fonttools==4.53.1
fqdn==1.5.1
fsspec==2024.9.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.67.0
h11==0.14.0
h5py==3.12.1
html5lib==1.1
httpcore==1.0.4
httpx==0.27.0
idna==3.8
ipykernel==6.29.3
ipython==8.22.2
ipywidgets==8.1.2
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
json5==0.9.24
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
jupyterlab_widgets==3.0.10
keras==3.6.0
kiwisolver==1.4.7
libclang==18.1.1
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.1
mpmath==1.3.0
mysqlclient==2.2.4
namex==0.0.8
narwhals==1.13.1
nbclient==0.10.0
nbconvert==7.16.2
nbformat==5.10.3
nest-asyncio==1.6.0
networkx==3.4.1
nltk==3.9.1
notebook==7.1.2
notebook_shim==0.2.4
numpy==1.26.4
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.13.0
oscrypto==1.3.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
panda==0.3.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.3
patsy==0.5.6
pillow==10.4.0
platformdirs==4.2.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.5
psutil==5.9.8
pure-eval==0.2.2
pyarrow==18.0.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.17.2
PyJWT==2.10.1
pymongo==4.11
pyparsing==3.1.4
pypdf==4.3.1
pypng==0.20220715.0
PySocks==1.7.1
pyTelegramBotAPI==4.26.0
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-telegram-bot==21.10
pytz==2024.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.2
pyzmq==25.1.2
qrcode==7.4.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.34.0
regex==2024.7.24
reportlab==4.2.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.2
rpds-py==0.18.0
scikit-learn==1.5.1
scipy==1.14.1
seaborn==0.13.2
selenium==4.27.1
Send2Trash==1.8.2
shortuuid==1.0.13
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.45.3
statsmodels==0.14.2
streamlit==1.39.0
svglib==1.5.1
sympy==1.13.1
telebot==0.0.5
tenacity==9.0.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.3.0
toml==0.10.2
tomli==2.0.1
torch==2.5.0
tornado==6.4
tqdm==4.66.5
traitlets==5.14.2
trio==0.27.0
trio-websocket==0.11.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uri-template==1.3.0
uritools==4.0.3
urllib3==2.2.2
uvicorn==0.34.0
validate-email==1.3
watchdog==5.0.3
wcwidth==0.2.13
webcolors==1.13
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.10
wrapt==1.16.0
wsproto==1.2.0
xgboost==1.5.0
